+++ src/main/java/com/hbm/blocks/machine/MachineReactorBreeding.java
@@ -1,0 +1,64 @@
+package com.hbm.blocks.machine;
+
+
+import com.hbm.blocks.BlockDummyable;
+import com.hbm.blocks.ModBlocks;
+import com.hbm.interfaces.IMultiblock;
+import com.hbm.main.MainRegistry;
+import com.hbm.tileentity.TileEntityProxyInventory;
+import com.hbm.tileentity.machine.TileEntityMachineReactorBreeding;
+
+import cpw.mods.fml.common.network.internal.FMLNetworkHandler;
+import net.minecraft.block.material.Material;
+import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.tileentity.TileEntity;
+import net.minecraft.world.World;
+
+public class MachineReactorBreeding extends BlockDummyable implements IMultiblock {
+	
+	public MachineReactorBreeding(Material mat) {
+		super(mat);
+	}
+
+	@Override
+	public TileEntity createNewTileEntity(World world, int meta) {
+		
+		if(meta >= 12)
+			return new TileEntityMachineReactorBreeding();
+		
+		return new TileEntityProxyInventory();
+	}
+	
+	@Override
+	public boolean onBlockActivated(World world, int x, int y, int z, EntityPlayer player, int side, float hitX, float hitY, float hitZ) {
+		if(world.isRemote)
+		{
+			return true;
+		} else if(!player.isSneaking())
+		{
+			int[] pos = this.findCore(world, x, y, z);
+			
+			if(pos == null)
+				return false;
+			
+			TileEntityMachineReactorBreeding entity = (TileEntityMachineReactorBreeding) world.getTileEntity(pos[0], pos[1], pos[2]);
+			if(entity != null)
+			{
+				FMLNetworkHandler.openGui(player, MainRegistry.instance, ModBlocks.guiID_reactor_breeding, world, pos[0], pos[1], pos[2]);
+			}
+			return true;
+		} else {
+			return false;
+		}
+	}
+
+	@Override
+	public int[] getDimensions() {
+		return new int[] { 2, 0, 0, 0, 0, 0 };
+	}
+
+	@Override
+	public int getOffset() {
+		return 0;
+	}
+}
