+++ src/main/java/com/hbm/inventory/container/ContainerReactorResearch.java
@@ -1,34 +1,34 @@
 package com.hbm.inventory.container;
 
 import com.hbm.inventory.SlotMachineOutput;
-import com.hbm.tileentity.machine.TileEntityMachineReactorSmall;
+import com.hbm.tileentity.machine.TileEntityReactorResearch;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.player.InventoryPlayer;
 import net.minecraft.inventory.Container;
 import net.minecraft.inventory.Slot;
 import net.minecraft.item.ItemStack;
 
-public class ContainerMachineReactorSmall extends Container {
+public class ContainerReactorResearch extends Container {
 
-private TileEntityMachineReactorSmall seleniumEngine;
+private TileEntityReactorResearch reactor;
 	
-	public ContainerMachineReactorSmall(InventoryPlayer invPlayer, TileEntityMachineReactorSmall tedf) {
+	public ContainerReactorResearch(InventoryPlayer invPlayer, TileEntityReactorResearch tedf) {
 		
-		seleniumEngine = tedf;
+		reactor = tedf;
 		
 		//Rods
-		this.addSlotToContainer(new Slot(tedf, 0, 98, 18));
-		this.addSlotToContainer(new Slot(tedf, 1, 134, 18));
-		this.addSlotToContainer(new Slot(tedf, 2, 80, 36));
-		this.addSlotToContainer(new Slot(tedf, 3, 116, 36));
-		this.addSlotToContainer(new Slot(tedf, 4, 152, 36));
-		this.addSlotToContainer(new Slot(tedf, 5, 98, 54));
-		this.addSlotToContainer(new Slot(tedf, 6, 134, 54));
-		this.addSlotToContainer(new Slot(tedf, 7, 80, 72));
-		this.addSlotToContainer(new Slot(tedf, 8, 116, 72));
-		this.addSlotToContainer(new Slot(tedf, 9, 152, 72));
-		this.addSlotToContainer(new Slot(tedf, 10, 98, 90));
-		this.addSlotToContainer(new Slot(tedf, 11, 134, 90));
+		this.addSlotToContainer(new Slot(tedf, 0, 95, 22));
+		this.addSlotToContainer(new Slot(tedf, 1, 131, 22));
+		this.addSlotToContainer(new Slot(tedf, 2, 77, 40));
+		this.addSlotToContainer(new Slot(tedf, 3, 112, 40));
+		this.addSlotToContainer(new Slot(tedf, 4, 149, 40));
+		this.addSlotToContainer(new Slot(tedf, 5, 95, 58));
+		this.addSlotToContainer(new Slot(tedf, 6, 131, 58));
+		this.addSlotToContainer(new Slot(tedf, 7, 77, 76));
+		this.addSlotToContainer(new Slot(tedf, 8, 112, 76));
+		this.addSlotToContainer(new Slot(tedf, 9, 149, 76));
+		this.addSlotToContainer(new Slot(tedf, 10, 95, 94));
+		this.addSlotToContainer(new Slot(tedf, 11, 131, 94));
 		
 		for(int i = 0; i < 3; i++)
 		{
@@ -75,6 +75,6 @@ private TileEntityMachineReactorSmall seleniumEngine;
 
 	@Override
 	public boolean canInteractWith(EntityPlayer player) {
-		return seleniumEngine.isUseableByPlayer(player);
+		return reactor.isUseableByPlayer(player);
 	}
 }
