+++ src/main/java/com/hbm/inventory/gui/GUIReactorResearch.java
@@ -5,11 +5,11 @@ import org.lwjgl.input.Keyboard;
 import org.lwjgl.opengl.GL11;
 
 import com.hbm.inventory.FluidTank;
-import com.hbm.inventory.container.ContainerMachineReactorSmall;
+import com.hbm.inventory.container.ContainerReactorResearch;
 import com.hbm.lib.RefStrings;
 import com.hbm.packet.NBTControlPacket;
 import com.hbm.packet.PacketDispatcher;
-import com.hbm.tileentity.machine.TileEntityMachineReactorSmall;
+import com.hbm.tileentity.machine.TileEntityReactorResearch;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.audio.PositionedSoundRecord;
 import net.minecraft.client.gui.GuiTextField;
@@ -20,21 +20,23 @@ import net.minecraft.util.EnumChatFormatting;
 import net.minecraft.util.MathHelper;
 import net.minecraft.util.ResourceLocation;
 
-public class GUIMachineReactorSmall extends GuiInfoContainer {
+public class GUIReactorResearch extends GuiInfoContainer {
 
-	private static ResourceLocation texture = new ResourceLocation(RefStrings.MODID + ":textures/gui/gui_reactor_experimental.png");
-	private TileEntityMachineReactorSmall reactor;
-	private final NumberDisplay[] displays = new NumberDisplay[2];
+	private static ResourceLocation texture = new ResourceLocation(RefStrings.MODID + ":textures/gui/reactors/gui_research_reactor.png");
+	private TileEntityReactorResearch reactor;
+	private final NumberDisplay[] displays = new NumberDisplay[3];
+	byte timer;
 	
 	private GuiTextField field;
 
-	public GUIMachineReactorSmall(InventoryPlayer invPlayer, TileEntityMachineReactorSmall te) {
-		super(new ContainerMachineReactorSmall(invPlayer, te));
+	public GUIReactorResearch(InventoryPlayer invPlayer, TileEntityReactorResearch te) {
+		super(new ContainerReactorResearch(invPlayer, te));
 		reactor = te;
 		this.xSize = 176;
 		this.ySize = 222;
-		displays[0] = new NumberDisplay(12, 19).setDigitLength(4);
-		displays[1] = new NumberDisplay(12, 55).setDigitLength(3);
+		displays[0] = new NumberDisplay(14, 25).setDigitLength(4);
+		displays[1] = new NumberDisplay(12, 63).setDigitLength(3);
+		displays[2] = new NumberDisplay(5, 101).setDigitLength(3);
 	}
 	
 	@Override
@@ -46,9 +48,7 @@ public class GUIMachineReactorSmall extends GuiInfoContainer {
 		
 		Keyboard.enableRepeatEvents(true);
 		
-		this.field = new GuiTextField(this.fontRendererObj, guiLeft + 11, guiTop + 86, 35, 9);
-		this.field.setTextColor(0x00ff00);
-		this.field.setDisabledTextColour(0x008000);
+		this.field = new GuiTextField(this.fontRendererObj, guiLeft + 8, guiTop + 99, 33, 16);
 		this.field.setEnableBackgroundDrawing(false);
 		this.field.setMaxStringLength(3);
 	}
@@ -56,45 +56,33 @@ public class GUIMachineReactorSmall extends GuiInfoContainer {
 	@Override
 	public void drawScreen(int mouseX, int mouseY, float f) {
 		super.drawScreen(mouseX, mouseY, f);
-		
-		this.drawCustomInfo(this, mouseX, mouseY, guiLeft + 80, guiTop + 120, 88, 4, new String[] { "Core Temperature:", "   " + Math.round((reactor.heat) * 0.00002 * 980 + 20) + "°C" });
-		
-		String[] text = new String[] { "Coolant will move heat from the core to",
-				"the hull. Water will use that heat and",
-				"generate steam.",
-				"Water consumption rate:",
-				" 100 mB/t",
-				" 2000 mB/s",
-				"Coolant consumption rate:",
-				" 10 mB/t",
-				" 200 mB/s",
-				"Water next to the reactor's open",
-				"sides will pour into the tank." };
-		this.drawCustomInfoStat(mouseX, mouseY, guiLeft - 16, guiTop + 36, 16, 16, guiLeft - 8, guiTop + 36 + 16, text);
-		
-		String[] text1 = new String[] { "Raise/lower the control rods",
-				"using the button next to the",
-				"fluid gauges." };
-		this.drawCustomInfoStat(mouseX, mouseY, guiLeft - 16, guiTop + 36 + 16, 16, 16, guiLeft - 8, guiTop + 36 + 16, text1);
+				
+		String[] text = new String[] {
+				"The reactor has to be submerged",
+				"in water on its sides to cool.",
+				"The neutron flux is provided to",
+				"adjacent breeding reactors."
+		};
+		this.drawCustomInfoStat(mouseX, mouseY, guiLeft - 14, guiTop + 23, 16, 16, guiLeft - 6, guiTop + 23 + 16, text);
 	}
 	
 	@Override
 	protected void drawGuiContainerForegroundLayer(int i, int j) {
 		String name = this.reactor.hasCustomInventoryName() ? this.reactor.getInventoryName() : I18n.format(this.reactor.getInventoryName());
-		final String[] labels = { "Flux", "Heat", "Control Rods" };
+		final String[] labels = { "Flux", "Heat", "Control" };
 		
-		this.fontRendererObj.drawString(name, 124 - this.fontRendererObj.getStringWidth(name) / 2, 6, 4210752);
+		this.fontRendererObj.drawString(name, 121 - this.fontRendererObj.getStringWidth(name) / 2, 6, 15066597);
 		this.fontRendererObj.drawString(I18n.format("container.inventory"), 8, this.ySize - 96 + 2, 4210752);
-		this.fontRendererObj.drawString(labels[0], 11, 9, 15066597);
-		this.fontRendererObj.drawString(labels[1], 11, 45, 15066597);
-		this.fontRendererObj.drawString(labels[2], 9, 74, 4210752);
+		this.fontRendererObj.drawString(labels[0], 6, 13, 15066597);
+		this.fontRendererObj.drawString(labels[1], 6, 51, 15066597);
+		this.fontRendererObj.drawString(labels[2], 6, 89, 15066597);
 	}
 
     protected void mouseClicked(int mouseX, int mouseY, int i) {
     	super.mouseClicked(mouseX, mouseY, i);
     	this.field.mouseClicked(mouseX, mouseY, i);
     	
-    	if(guiLeft + 51 <= mouseX && guiLeft + 51 + 12 > mouseX && guiTop + 84 < mouseY && guiTop + 84 + 12 >= mouseY) {
+    	if(guiLeft + 44 <= mouseX && guiLeft + 44 + 11 > mouseX && guiTop + 97 < mouseY && guiTop + 97 + 20 >= mouseY) {
 			
     		double level;
     		
@@ -108,6 +96,7 @@ public class GUIMachineReactorSmall extends GuiInfoContainer {
 			
 			NBTTagCompound control = new NBTTagCompound();
 			control.setDouble("level", level);
+			timer = 15;
 			
 			PacketDispatcher.wrapper.sendToServer(new NBTControlPacket(control, reactor.xCoord, reactor.yCoord, reactor.zCoord));
 			mc.getSoundHandler().playSound(PositionedSoundRecord.func_147674_a(new ResourceLocation("hbm:block.rbmk_az5_cover"), 0.5F));
@@ -121,33 +110,31 @@ public class GUIMachineReactorSmall extends GuiInfoContainer {
 		
 		drawTexturedModalRect(guiLeft, guiTop, 0, 0, xSize, ySize);
 		
-		if(reactor.heat > 0) {
-			int i = reactor.getHeatScaled(88);
-			
-			i = (int) Math.min(i, 160);
-			
-			drawTexturedModalRect(guiLeft + 80, guiTop + 120, 0, 230, i, 4);
+		if(reactor.level <= 0.5D) {
+			for(int x = 0; x < 3; x++)
+				for(int y = 0; y < 3; y++)
+					drawTexturedModalRect(guiLeft + 81 + 36 * x, guiTop + 26 + 36 * y, 176, 0, 8, 8);
 		}
 		
-		if(reactor.level >= 100) {	
-			for(int x = 0; x < 3; x++)
-				for(int y = 0; y < 3; y++)
-					drawTexturedModalRect(guiLeft + 79 + 36 * x, guiTop + 17 + 36 * y, 176, 0, 18, 18);
-				
-		} else if(reactor.level > 0) {
-			for(int x = 0; x < 3; x++)
-				for(int y = 0; y < 3; y++)
-					drawTexturedModalRect(guiLeft + 79 + 36 * x, guiTop + 17 + 36 * y, 194, 0, 18, 18);
-				
+		if(timer > 0) {
+			drawTexturedModalRect(guiLeft + 44, guiTop + 97, 176, 8, 11, 20);
+			timer--;
 		}
 		
 		for(byte i = 0; i < 2; i++)
 			displays[i].drawNumber(reactor.getDisplayData()[i]);
 		
-		this.drawInfoPanel(guiLeft - 16, guiTop + 36, 16, 16, 2);
-		this.drawInfoPanel(guiLeft - 16, guiTop + 36 + 16, 16, 16, 3);
+		if(NumberUtils.isDigits(field.getText())) {
+			int level = (int)MathHelper.clamp_double(Double.parseDouble(field.getText()), 0, 100);
+			field.setText(level + "");
+			displays[2].drawNumber(level);
+		} else {
+			field.setText(0 + "");
+			displays[2].drawNumber(0);
+		}
 		
-		this.field.drawTextBox();
+		this.drawInfoPanel(guiLeft - 14, guiTop + 23, 16, 16, 3);
+		
 	}
 	
 	@Override
