+++ src/main/java/com/hbm/inventory/recipes/BreederRecipes.java
@@ -18,120 +18,64 @@ import net.minecraft.util.EnumChatFormatting;
 public class BreederRecipes {
 
 	private static HashMap<ComparableStack, BreederRecipe> recipes = new HashMap();
-	private static HashMap<ComparableStack, int[]> fuels = new HashMap();
-	//for the int array: [0] => level (1-4) [1] => amount of operations
 	
 	public static void registerRecipes() {
 
 		//lithium and impure rods
-		recipes.put(new ComparableStack(ModItems.rod_lithium), new BreederRecipe(ModItems.rod_tritium, 1));
-		recipes.put(new ComparableStack(ModItems.rod_dual_lithium), new BreederRecipe(ModItems.rod_dual_tritium, 1));
-		recipes.put(new ComparableStack(ModItems.rod_quad_lithium), new BreederRecipe(ModItems.rod_quad_tritium, 1));
-		recipes.put(new ComparableStack(ModItems.rod_uranium), new BreederRecipe(ModItems.rod_plutonium, 4));
-		recipes.put(new ComparableStack(ModItems.rod_dual_uranium), new BreederRecipe(ModItems.rod_dual_plutonium, 4));
-		recipes.put(new ComparableStack(ModItems.rod_quad_uranium), new BreederRecipe(ModItems.rod_quad_plutonium, 4));
-		recipes.put(new ComparableStack(ModItems.rod_plutonium), new BreederRecipe(ModItems.rod_waste, 4));
-		recipes.put(new ComparableStack(ModItems.rod_dual_plutonium), new BreederRecipe(ModItems.rod_dual_waste, 4));
-		recipes.put(new ComparableStack(ModItems.rod_quad_plutonium), new BreederRecipe(ModItems.rod_quad_waste, 4));
+		recipes.put(new ComparableStack(ModItems.rod_lithium), new BreederRecipe(ModItems.rod_tritium, 100));
+		recipes.put(new ComparableStack(ModItems.rod_dual_lithium), new BreederRecipe(ModItems.rod_dual_tritium, 200));
+		recipes.put(new ComparableStack(ModItems.rod_quad_lithium), new BreederRecipe(ModItems.rod_quad_tritium, 400));
+		recipes.put(new ComparableStack(ModItems.rod_uranium), new BreederRecipe(ModItems.rod_plutonium, 1000));
+		recipes.put(new ComparableStack(ModItems.rod_dual_uranium), new BreederRecipe(ModItems.rod_dual_plutonium, 2000));
+		recipes.put(new ComparableStack(ModItems.rod_quad_uranium), new BreederRecipe(ModItems.rod_quad_plutonium, 4000));
+		recipes.put(new ComparableStack(ModItems.rod_plutonium), new BreederRecipe(ModItems.rod_waste, 500));
+		recipes.put(new ComparableStack(ModItems.rod_dual_plutonium), new BreederRecipe(ModItems.rod_dual_waste, 1000));
+		recipes.put(new ComparableStack(ModItems.rod_quad_plutonium), new BreederRecipe(ModItems.rod_quad_waste, 2000));
 		
 		//isotopes
-		recipes.put(new ComparableStack(ModItems.rod_th232), new BreederRecipe(ModItems.rod_u233, 2));
-		recipes.put(new ComparableStack(ModItems.rod_dual_th232), new BreederRecipe(ModItems.rod_dual_u233, 2));
-		recipes.put(new ComparableStack(ModItems.rod_quad_th232), new BreederRecipe(ModItems.rod_quad_u233, 2));
-		recipes.put(new ComparableStack(ModItems.rod_u233), new BreederRecipe(ModItems.rod_u235, 2));
-		recipes.put(new ComparableStack(ModItems.rod_dual_u233), new BreederRecipe(ModItems.rod_dual_u235, 2));
-		recipes.put(new ComparableStack(ModItems.rod_quad_u233), new BreederRecipe(ModItems.rod_quad_u235, 2));
-		recipes.put(new ComparableStack(ModItems.rod_u235), new BreederRecipe(ModItems.rod_neptunium, 3));
-		recipes.put(new ComparableStack(ModItems.rod_dual_u235), new BreederRecipe(ModItems.rod_dual_neptunium, 3));
-		recipes.put(new ComparableStack(ModItems.rod_quad_u235), new BreederRecipe(ModItems.rod_quad_neptunium, 3));
-		recipes.put(new ComparableStack(ModItems.rod_u238), new BreederRecipe(ModItems.rod_pu239, 3));
-		recipes.put(new ComparableStack(ModItems.rod_dual_u238), new BreederRecipe(ModItems.rod_dual_pu239, 3));
-		recipes.put(new ComparableStack(ModItems.rod_quad_u238), new BreederRecipe(ModItems.rod_quad_pu239, 3));
-		recipes.put(new ComparableStack(ModItems.rod_neptunium), new BreederRecipe(ModItems.rod_pu238, 3));
-		recipes.put(new ComparableStack(ModItems.rod_dual_neptunium), new BreederRecipe(ModItems.rod_dual_pu238, 3));
-		recipes.put(new ComparableStack(ModItems.rod_quad_neptunium), new BreederRecipe(ModItems.rod_quad_pu238, 3));
-		recipes.put(new ComparableStack(ModItems.rod_pu238), new BreederRecipe(ModItems.rod_pu239, 4));
-		recipes.put(new ComparableStack(ModItems.rod_dual_pu238), new BreederRecipe(ModItems.rod_dual_pu239, 4));
-		recipes.put(new ComparableStack(ModItems.rod_quad_pu238), new BreederRecipe(ModItems.rod_quad_pu239, 4));
-		recipes.put(new ComparableStack(ModItems.rod_pu239), new BreederRecipe(ModItems.rod_pu240, 2));
-		recipes.put(new ComparableStack(ModItems.rod_dual_pu239), new BreederRecipe(ModItems.rod_dual_pu240, 2));
-		recipes.put(new ComparableStack(ModItems.rod_quad_pu239), new BreederRecipe(ModItems.rod_quad_pu240, 2));
-		recipes.put(new ComparableStack(ModItems.rod_pu240), new BreederRecipe(ModItems.rod_waste, 3));
-		recipes.put(new ComparableStack(ModItems.rod_dual_pu240), new BreederRecipe(ModItems.rod_dual_waste, 3));
-		recipes.put(new ComparableStack(ModItems.rod_quad_pu240), new BreederRecipe(ModItems.rod_quad_waste, 3));
+		recipes.put(new ComparableStack(ModItems.rod_th232), new BreederRecipe(ModItems.rod_u233, 1000));
+		recipes.put(new ComparableStack(ModItems.rod_dual_th232), new BreederRecipe(ModItems.rod_dual_u233, 2000));
+		recipes.put(new ComparableStack(ModItems.rod_quad_th232), new BreederRecipe(ModItems.rod_quad_u233, 4000));
+		recipes.put(new ComparableStack(ModItems.rod_u233), new BreederRecipe(ModItems.rod_u235, 1000));
+		recipes.put(new ComparableStack(ModItems.rod_dual_u233), new BreederRecipe(ModItems.rod_dual_u235, 2000));
+		recipes.put(new ComparableStack(ModItems.rod_quad_u233), new BreederRecipe(ModItems.rod_quad_u235, 4000));
+		recipes.put(new ComparableStack(ModItems.rod_u235), new BreederRecipe(ModItems.rod_neptunium, 500));
+		recipes.put(new ComparableStack(ModItems.rod_dual_u235), new BreederRecipe(ModItems.rod_dual_neptunium, 1000));
+		recipes.put(new ComparableStack(ModItems.rod_quad_u235), new BreederRecipe(ModItems.rod_quad_neptunium, 2000));
+		recipes.put(new ComparableStack(ModItems.rod_u238), new BreederRecipe(ModItems.rod_pu239, 1000));
+		recipes.put(new ComparableStack(ModItems.rod_dual_u238), new BreederRecipe(ModItems.rod_dual_pu239, 2000));
+		recipes.put(new ComparableStack(ModItems.rod_quad_u238), new BreederRecipe(ModItems.rod_quad_pu239, 4000));
+		recipes.put(new ComparableStack(ModItems.rod_neptunium), new BreederRecipe(ModItems.rod_pu238, 250));
+		recipes.put(new ComparableStack(ModItems.rod_dual_neptunium), new BreederRecipe(ModItems.rod_dual_pu238, 500));
+		recipes.put(new ComparableStack(ModItems.rod_quad_neptunium), new BreederRecipe(ModItems.rod_quad_pu238, 1000));
+		recipes.put(new ComparableStack(ModItems.rod_pu238), new BreederRecipe(ModItems.rod_pu239, 1000));
+		recipes.put(new ComparableStack(ModItems.rod_dual_pu238), new BreederRecipe(ModItems.rod_dual_pu239, 2000));
+		recipes.put(new ComparableStack(ModItems.rod_quad_pu238), new BreederRecipe(ModItems.rod_quad_pu239, 4000));
+		recipes.put(new ComparableStack(ModItems.rod_pu239), new BreederRecipe(ModItems.rod_pu240, 500));
+		recipes.put(new ComparableStack(ModItems.rod_dual_pu239), new BreederRecipe(ModItems.rod_dual_pu240, 1000));
+		recipes.put(new ComparableStack(ModItems.rod_quad_pu239), new BreederRecipe(ModItems.rod_quad_pu240, 2000));
+		recipes.put(new ComparableStack(ModItems.rod_pu240), new BreederRecipe(ModItems.rod_waste, 500));
+		recipes.put(new ComparableStack(ModItems.rod_dual_pu240), new BreederRecipe(ModItems.rod_dual_waste, 1000));
+		recipes.put(new ComparableStack(ModItems.rod_quad_pu240), new BreederRecipe(ModItems.rod_quad_waste, 2000));
 
 		//advanced
-		recipes.put(new ComparableStack(ModItems.rod_schrabidium), new BreederRecipe(ModItems.rod_solinium, 3));
-		recipes.put(new ComparableStack(ModItems.rod_dual_schrabidium), new BreederRecipe(ModItems.rod_dual_solinium, 3));
-		recipes.put(new ComparableStack(ModItems.rod_quad_schrabidium), new BreederRecipe(ModItems.rod_quad_solinium, 3));
-		recipes.put(new ComparableStack(ModItems.rod_quad_solinium), new BreederRecipe(ModItems.rod_quad_euphemium, 4));
-		recipes.put(new ComparableStack(ModItems.rod_balefire), new BreederRecipe(ModItems.rod_balefire_blazing, 4));
-		recipes.put(new ComparableStack(ModItems.rod_dual_balefire), new BreederRecipe(ModItems.rod_dual_balefire_blazing, 4));
-		recipes.put(new ComparableStack(ModItems.rod_quad_balefire), new BreederRecipe(ModItems.rod_quad_balefire_blazing, 4));
+		recipes.put(new ComparableStack(ModItems.rod_schrabidium), new BreederRecipe(ModItems.rod_solinium, 2000));
+		recipes.put(new ComparableStack(ModItems.rod_dual_schrabidium), new BreederRecipe(ModItems.rod_dual_solinium, 4000));
+		recipes.put(new ComparableStack(ModItems.rod_quad_schrabidium), new BreederRecipe(ModItems.rod_quad_solinium, 8000));
+		recipes.put(new ComparableStack(ModItems.rod_quad_solinium), new BreederRecipe(ModItems.rod_quad_euphemium, 2000));
+		recipes.put(new ComparableStack(ModItems.rod_balefire), new BreederRecipe(ModItems.rod_balefire_blazing, 2000));
+		recipes.put(new ComparableStack(ModItems.rod_dual_balefire), new BreederRecipe(ModItems.rod_dual_balefire_blazing, 4000));
+		recipes.put(new ComparableStack(ModItems.rod_quad_balefire), new BreederRecipe(ModItems.rod_quad_balefire_blazing, 8000));
 
 		//rocks
-		recipes.put(new ComparableStack(Blocks.stone), new BreederRecipe(new ItemStack(ModBlocks.sellafield_0), 2));
-		recipes.put(new ComparableStack(ModBlocks.sellafield_0), new BreederRecipe(new ItemStack(ModBlocks.sellafield_1), 2));
-		recipes.put(new ComparableStack(ModBlocks.sellafield_1), new BreederRecipe(new ItemStack(ModBlocks.sellafield_2), 3));
-		recipes.put(new ComparableStack(ModBlocks.sellafield_2), new BreederRecipe(new ItemStack(ModBlocks.sellafield_3), 3));
-		recipes.put(new ComparableStack(ModBlocks.sellafield_3), new BreederRecipe(new ItemStack(ModBlocks.sellafield_4), 4));
-		recipes.put(new ComparableStack(ModBlocks.sellafield_4), new BreederRecipe(new ItemStack(ModBlocks.sellafield_core), 4));
+		recipes.put(new ComparableStack(Blocks.stone), new BreederRecipe(new ItemStack(ModBlocks.sellafield_0), 250));
+		recipes.put(new ComparableStack(ModBlocks.sellafield_0), new BreederRecipe(new ItemStack(ModBlocks.sellafield_1), 250));
+		recipes.put(new ComparableStack(ModBlocks.sellafield_1), new BreederRecipe(new ItemStack(ModBlocks.sellafield_2), 500));
+		recipes.put(new ComparableStack(ModBlocks.sellafield_2), new BreederRecipe(new ItemStack(ModBlocks.sellafield_3), 500));
+		recipes.put(new ComparableStack(ModBlocks.sellafield_3), new BreederRecipe(new ItemStack(ModBlocks.sellafield_4), 1000));
+		recipes.put(new ComparableStack(ModBlocks.sellafield_4), new BreederRecipe(new ItemStack(ModBlocks.sellafield_core), 1000));
 		
-		recipes.put(new ComparableStack(ModItems.meteorite_sword_etched), new BreederRecipe(new ItemStack(ModItems.meteorite_sword_bred), 4));
-	}
-	
-	public static void registerFuels() {
-		fuels.put(new ComparableStack(ModItems.rod_u233), new int[] {2, 2});
-		fuels.put(new ComparableStack(ModItems.rod_dual_u233), new int[] {2, 4});
-		fuels.put(new ComparableStack(ModItems.rod_quad_u233), new int[] {2, 8});
-		
-		fuels.put(new ComparableStack(ModItems.rod_u235), new int[] {2, 3});
-		fuels.put(new ComparableStack(ModItems.rod_dual_u235), new int[] {2, 6});
-		fuels.put(new ComparableStack(ModItems.rod_quad_u235), new int[] {2, 12});
-		
-		fuels.put(new ComparableStack(ModItems.rod_u238), new int[] {1, 1});
-		fuels.put(new ComparableStack(ModItems.rod_dual_u238), new int[] {1, 2});
-		fuels.put(new ComparableStack(ModItems.rod_quad_u238), new int[] {1, 4});
-		
-		fuels.put(new ComparableStack(ModItems.rod_neptunium), new int[] {2, 3});
-		fuels.put(new ComparableStack(ModItems.rod_dual_neptunium), new int[] {2, 6});
-		fuels.put(new ComparableStack(ModItems.rod_quad_neptunium), new int[] {2, 12});
-		
-		fuels.put(new ComparableStack(ModItems.rod_pu238), new int[] {1, 2});
-		fuels.put(new ComparableStack(ModItems.rod_dual_pu238), new int[] {1, 4});
-		fuels.put(new ComparableStack(ModItems.rod_quad_pu238), new int[] {1, 8});
-		
-		fuels.put(new ComparableStack(ModItems.rod_pu239), new int[] {3, 5});
-		fuels.put(new ComparableStack(ModItems.rod_dual_pu239), new int[] {3, 10});
-		fuels.put(new ComparableStack(ModItems.rod_quad_pu239), new int[] {3, 20});
-		
-		fuels.put(new ComparableStack(ModItems.rod_pu240), new int[] {1, 2});
-		fuels.put(new ComparableStack(ModItems.rod_dual_pu240), new int[] {1, 4});
-		fuels.put(new ComparableStack(ModItems.rod_quad_pu240), new int[] {1, 8});
-		
-		fuels.put(new ComparableStack(ModItems.rod_schrabidium), new int[] {3, 10});
-		fuels.put(new ComparableStack(ModItems.rod_dual_schrabidium), new int[] {3, 20});
-		fuels.put(new ComparableStack(ModItems.rod_quad_schrabidium), new int[] {3, 40});
-		
-		fuels.put(new ComparableStack(ModItems.rod_solinium), new int[] {3, 15});
-		fuels.put(new ComparableStack(ModItems.rod_dual_solinium), new int[] {3, 30});
-		fuels.put(new ComparableStack(ModItems.rod_quad_solinium), new int[] {3, 60});
-		
-		fuels.put(new ComparableStack(ModItems.rod_polonium), new int[] {4, 2});
-		fuels.put(new ComparableStack(ModItems.rod_dual_polonium), new int[] {4, 4});
-		fuels.put(new ComparableStack(ModItems.rod_quad_polonium), new int[] {4, 8});
-		
-		fuels.put(new ComparableStack(ModItems.rod_tritium), new int[] {1, 1});
-		fuels.put(new ComparableStack(ModItems.rod_dual_tritium), new int[] {1, 2});
-		fuels.put(new ComparableStack(ModItems.rod_quad_tritium), new int[] {1, 4});
-		
-		fuels.put(new ComparableStack(ModItems.rod_balefire), new int[] {2, 150});
-		fuels.put(new ComparableStack(ModItems.rod_dual_balefire), new int[] {2, 300});
-		fuels.put(new ComparableStack(ModItems.rod_quad_balefire), new int[] {2, 600});
-		
-		fuels.put(new ComparableStack(ModItems.rod_balefire_blazing), new int[] {4, 75});
-		fuels.put(new ComparableStack(ModItems.rod_dual_balefire_blazing), new int[] {4, 150});
-		fuels.put(new ComparableStack(ModItems.rod_quad_balefire_blazing), new int[] {4, 300});
+		recipes.put(new ComparableStack(ModItems.meteorite_sword_etched), new BreederRecipe(new ItemStack(ModItems.meteorite_sword_bred), 1000));
 	}
 	
 	public static HashMap<ItemStack, BreederRecipe> getAllRecipes() {
@@ -145,20 +89,6 @@ public class BreederRecipes {
 		return map;
 	}
 	
-	public static List<ItemStack> getAllFuelsFromHEAT(int heat) {
-		
-		List<ItemStack> list = new ArrayList();
-		
-		for(Map.Entry<ComparableStack, int[]> fuel : fuels.entrySet()) {
-			
-			if(fuel.getValue()[0] >= heat) {
-				list.add(fuel.getKey().toStack());
-			}
-		}
-		
-		return list;
-	}
-	
 	public static BreederRecipe getOutput(ItemStack stack) {
 		
 		if(stack == null)
@@ -168,51 +98,19 @@ public class BreederRecipes {
 		return BreederRecipes.recipes.get(sta);
 	}
 	
-	/**
-	 * Returns an integer array of the fuel value of a certain stack
-	 * @param stack
-	 * @return an integer array (possibly null) with two fields, the HEAT value and the amount of operations
-	 */
-	public static int[] getFuelValue(ItemStack stack) {
-		
-		if(stack == null)
-			return null;
-		
-		ComparableStack sta = new ComparableStack(stack);
-		int[] ret = BreederRecipes.fuels.get(sta);
-		
-		return ret;
-	}
-	
-	public static String getHEATString(String string, int heat) {
-
-		if(heat == 1)
-			string =  EnumChatFormatting.GREEN + string;
-		if(heat == 2)
-			string = EnumChatFormatting.YELLOW + string;
-		if(heat == 3)
-			string = EnumChatFormatting.GOLD + string;
-		if(heat == 4)
-			string = EnumChatFormatting.RED + string;
-		
-		return string; //strings are reference types I GET IT
-	}
-	
 	//nicer than opaque object arrays
 	public static class BreederRecipe {
 		
 		public ItemStack output;
-		public int heat;
+		public int flux;
 		
-		public BreederRecipe() { }
-		
-		public BreederRecipe(Item output, int heat) {
-			this(new ItemStack(output), heat);
+		public BreederRecipe(Item output, int flux) {
+			this(new ItemStack(output), flux);
 		}
 		
-		public BreederRecipe(ItemStack output, int heat) {
+		public BreederRecipe(ItemStack output, int flux) {
 			this.output = output;
-			this.heat = heat;
+			this.flux = flux;
 		}
 	}
