+++ src/main/java/com/hbm/inventory/recipes/anvil/AnvilRecipes.java
@@ -173,7 +173,8 @@ public class AnvilRecipes {
 						new OreDictStack(POLYMER.ingot(), 4),
 						new ComparableStack(ModItems.generator_steel, 2),
 						new ComparableStack(ModItems.turbine_titanium, 1),
-						new ComparableStack(ModItems.thermo_element, 3)
+						new ComparableStack(ModItems.thermo_element, 3),
+						new ComparableStack(ModItems.crt_display, 1)
 				}, new AnvilOutput(new ItemStack(ModBlocks.machine_industrial_generator))).setTier(2));
 
 		constructionRecipes.add(new AnvilConstructionRecipe(
@@ -586,7 +587,7 @@ public class AnvilRecipes {
 							new AnvilOutput(new ItemStack(ModItems.pipes_steel, 2))
 					}).setTier(4));
 			constructionRecipes.add(new AnvilConstructionRecipe(
-					new ComparableStack(ModBlocks.machine_reactor_small), new AnvilOutput[] {
+					new ComparableStack(ModBlocks.reactor_research), new AnvilOutput[] {
 							new AnvilOutput(new ItemStack(ModItems.ingot_steel, 6)),
 							new AnvilOutput(new ItemStack(ModItems.ingot_polymer, 1)),
 							new AnvilOutput(new ItemStack(ModItems.ingot_polymer, 1), 0.75F),
