+++ src/main/resources/assets/hbm/lang/en_US.lang
@@ -383,10 +383,10 @@ container.rbmkOutgasser=RBMK Irradiation Channel
 container.rbmkReaSim=RBMK Fuel Rod (ReaSim)
 container.rbmkRod=RBMK Fuel Rod
 container.rbmkStorage=RBMK Storage Column
-container.reactor=Breeding Reactor
+container.reactorBreeding=Breeding Reactor
 container.reactorControl=Reactor Remote Control Block
 container.reactorLarge=Big Nuclear Reactor
-container.reactorSmall=Nuclear Reactor
+container.reactorResearch=Research Reactor
 container.reix=Rei-X Mainframe
 container.rtg=RT Generator
 container.rtgFurnace=RTG Furnace
@@ -1273,6 +1273,7 @@ item.crystal_trixite.name=Trixite Crystals
 item.crystal_tungsten.name=Tungsten Crystals
 item.crystal_uranium.name=Uranium Crystals
 item.crystal_xen.name=Artificial Xen Crystal
+item.crt_display.name= Cathode Ray Tube
 item.cube_power.name=Electronium Cube
 item.custom_amat.name=Custom Nuke Antimatter Rod
 item.custom_dirty.name=Custom Nuke Dirty Rod
@@ -3328,9 +3329,7 @@ tile.machine_puf6_tank.name=Plutonium Hexafluoride Tank
 tile.machine_pumpjack.name=Pumpjack
 tile.machine_radar.name=Radar
 tile.machine_radgen.name=Radiation-Powered Engine
-tile.machine_reactor.name=Breeding Reactor
-tile.machine_reactor_on.name=Breeding Reactor
-tile.machine_reactor_small.name=Nuclear Reactor
+tile.machine_reactor_breeding.name=Breeding Reactor
 tile.machine_refinery.name=Oil Refinery
 tile.machine_reix_mainframe.name=Rei-X Mainframe (WIP)
 tile.machine_rtg_blue.name=Convection Generator
@@ -3367,7 +3366,6 @@ tile.machine_turbofan.name=Turbofan
 tile.machine_uf6_tank.name=Uranium Hexafluoride Tank
 tile.machine_waste_drum.name=Spent Fuel Pool Drum
 tile.machine_well.name=Oil Derrick
-tile.machine_zirnox.name=ZIRNOX Nuclear Reactor
 tile.marker_structure.name=Multiblock Structure Marker
 tile.meteor_battery.name=Starmetal Static Electricity Generator
 tile.meteor_brick.name=Meteor Bricks
@@ -3512,6 +3510,8 @@ tile.reactor_ejector.name=Reactor Waste Ejector
 tile.reactor_element.name=Reactor Chamber
 tile.reactor_hatch.name=Reactor Access Hatch
 tile.reactor_inserter.name=Reactor Fuel Inserter
+tile.reactor_research.name=Research Reactor
+tile.reactor_zirnox.name=ZIRNOX Nuclear Reactor
 tile.red_barrel.name=Explosive Barrel
 tile.red_cable.name=Red Copper Cable
 tile.red_connector.name=Electricity Connector
@@ -3634,7 +3634,6 @@ tile.yellow_barrel.name=Radioactive Barrel
 
 trait.asbestos=Asbestos
 trait.blinding=Blinding
-trait.breeding=Worth %s operations in breeding reactor
 trait.coal=Coal Dust
 trait.digamma=Digamma Radiation
 trait.drop=Dangerous Drop
